🔧 chore: ErrorType Enum을 error 폴더로 이동
