+++ src/main/java/sopt/org/motivooServer/domain/mission/exception/MissionException.java
@@ -1,11 +1,11 @@
 package sopt.org.motivooServer.domain.mission.exception;
 
-import sopt.org.motivooServer.global.common.exception.ErrorType;
+import sopt.org.motivooServer.global.common.error.ErrorResponse;
 import sopt.org.motivooServer.global.common.exception.BusinessException;
 
 public class MissionException extends BusinessException {
 
-	public MissionException(ErrorType errorType) {
+	public MissionException(ErrorResponse.ErrorType errorType) {
 		super(errorType);
 	}
 }
