+++ src/main/java/sopt/org/motivooServer/domain/user/exception/UserException.java
@@ -1,11 +1,11 @@
 package sopt.org.motivooServer.domain.user.exception;
 
-import sopt.org.motivooServer.global.common.exception.ErrorType;
+import sopt.org.motivooServer.global.common.error.ErrorResponse;
 import sopt.org.motivooServer.global.common.exception.BusinessException;
 
 public class UserException extends BusinessException {
 
-	public UserException(ErrorType errorType) {
+	public UserException(ErrorResponse.ErrorType errorType) {
 		super(errorType);
 	}
 }
