+++ src/main/java/sopt/org/motivooServer/global/common/error/ErrorResponse.java
@@ -2,7 +2,9 @@ package sopt.org.motivooServer.global.common.error;
 
 import org.springframework.http.HttpStatus;
 
-import sopt.org.motivooServer.global.common.exception.ErrorType;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 
 public record ErrorResponse(
 	int code,
@@ -24,5 +26,24 @@ public record ErrorResponse(
 			message,
 			false
 		);
+	}
+
+	@Getter
+	@AllArgsConstructor(access = AccessLevel.PRIVATE)
+	public enum ErrorType {
+
+		/**
+		 * 500 Internal Server Error
+		 */
+		INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "알 수 없는 서버 에러가 발생했습니다"),
+
+		;
+
+		private final HttpStatus httpStatus;
+		private final String message;
+
+		public int getHttpStatusCode() {
+			return httpStatus.value();
+		}
 	}
 }
