+++ src/main/java/sopt/org/motivooServer/global/common/exception/BusinessException.java
@@ -1,13 +1,14 @@
 package sopt.org.motivooServer.global.common.exception;
 
 import lombok.Getter;
+import sopt.org.motivooServer.global.common.error.ErrorResponse;
 
 @Getter
 public class BusinessException extends RuntimeException {
 
-	private final ErrorType errorType;
+	private final ErrorResponse.ErrorType errorType;
 
-	public BusinessException(ErrorType errorType) {
+	public BusinessException(ErrorResponse.ErrorType errorType) {
 		super(errorType.getMessage());
 		this.errorType = errorType;
 	}
