--- src/main/java/sopt/org/motivooServer/global/common/exception/ErrorType.java
@@ -1,26 +1,0 @@
-package sopt.org.motivooServer.global.common.exception;
-
-import org.springframework.http.HttpStatus;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public enum ErrorType {
-
-	/**
-	 * 500 Internal Server Error
-	 */
-	INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "알 수 없는 서버 에러가 발생했습니다"),
-
-	;
-
-	private final HttpStatus httpStatus;
-	private final String message;
-
-	public int getHttpStatusCode() {
-		return httpStatus.value();
-	}
-}
